requirements.txt
google-generativeai
python-dotenv
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2